BUG/MAJOR: mux-h2: don't try to decode a response HEADERS frame in idle state
Christopher found another issue in the H2 backend implementation that
results from a miss in the H2 spec: the processing of a HEADERS frame
is always permitted in IDLE state, but this doesn't make sense on the
response path! And here when facing such a frame, we try to decode it
while we didn't allocate any stream, so we end up trying to fill the
idle stream's buffer (read-only) and crash.

What we're doing here is that if we get a HEADERS frame in IDLE state
from a server, we terminate the connection with a PROTOCOL_ERROR. No
such transition seems to be permitted by the spec but it seems to be
the only sane solution.

This fix must be backported as far as 1.9. Note that in 2.0 and earlier
there's no h2_frame_check_vs_state() function, instead the check is
inlined in h2_process_demux().

(cherry picked from commit 57a1816faec652d9c7fbd9be48ce25c832be20d9)
[wt: moved the code into h2_process_demux()]
Signed-off-by: Willy Tarreau <w@1wt.eu>

(cherry picked from commit 85ee6e8343101f598a6e7d9149af37a6579ec6ff)
Signed-off-by: Willy Tarreau <w@1wt.eu>